Order details update is working. You could change the count of some Product in curt. Also main page has been fixed and a few more bugs.

+++ src/main/java/com/lider/BlockNoteWebApp/controller/MainController.java
@@ -151,7 +151,7 @@ public class MainController {
             orderdetails.setAmount(quanity*price);
         }
 
-        double am = order.getAmount();
+        Integer am = order.getAmount();
         order.setAmount(am + price);
 
         OrderRepo.save(order);
