+++ src/main/java/com/lider/BlockNoteWebApp/controller/ProductController.java
@@ -70,7 +70,7 @@ public class ProductController {
             orderdetails.setQuanity(quanity);
             orderdetails.setAmount(quanity*price);
         }
-        double am = order.getAmount();
+        Integer am = order.getAmount();
         order.setAmount(am + price);
 
         OrderRepo.save(order);
